Started adding platforming scripts
- Added BoxTestScene to test out the "Box"/ platforming section of the game
- Added Collision Check & Capability abstract classes
- Added Move capability and GroundCheck collisionCheck

Scripts are still not entirely functional, but the code is not broken. What's in the Move capability and GroundCheck collisionCheck is mostly an outline.

Capability is for handling things like player movement. They all require a BoxPlayerHandler, which will receive events from PlayerInput components and trigger events. Capabilities will subscribe to the needed events from BoxPlayerHandler.

CollisionChecks are middle-man ScriptableObjects that handle calculations on collisions (whether or not a collided surface can be considered "ground" or not). Collision calculations are called by BoxPlayerHandler.
This has been how I've handled collision checks for the past few projects, although I may change this in the future if I happen to find a more efficient process.