Added Jump capability
- Added Jump capability, with jump force, coyote time, jump buffer, variable height, etc.
- Added Spacebar as a keyboard binding for submit
- Added a trail material to visualise player movement / jump curves in BoxTestScene

The jump works using the "Submit" button.

Jump values have not yet been tweaked. They should probably be tweaked when the rest of the moveset is implemented, so that each move is balanced with one another. The jump values should also be readjusted when player size is decided.

I also opted out of a double jump as of currently, as I want to add a different functionality to pressing jump while midair.