Started adding Smash capability
- Added Smash.cs, a capability script that will cover the game's main attack

The attack will be similar to the smash ability in ScourgeBringer, but upon hitting enemies/objects it'll launch the player in the opposite direction. This should be the "main" or most volatile capability the player has access to, and the moveset will be built around controlling the momentum generated from this.
The other capabilities will be revised in order to function well with this in future commits.