Added NewlyMatchedObjects to OnMatchFoundEvent
When OnMatchFoundEventArgs are made, they now require an array of GridObjects that weren't previously matched.

This is done through a new list, matchedObjects, and a method AddToMatchedObjects() which checks for duplicates before adding to the list, and only adds the non-duplicates. You can then get the non-duplicates from the method.