Added ability to hide GridObjectBehaviours
This was done through a new event, OnHideEvent ( to avoid GridObjects storing whether or not they're hidden ).

This event is called by a method in GridObjects, which is used by GridDisplay in a method that affects the whole grid. All of this is triggered when MatchTimer runs out.

This is a temporary solution for the grid to hide itself while the game shifts to the 'Box' mode.